PM / devfreq: Add cache HW monitor governor
The cache HW monitor devfreq governor uses the hardware counters to
determine the load on the cache and the appropriate frequency needed to
support that load. This governor can be used in conjunction with the
cache devfreq device to dynamically scale the cache frequency based on
the demand/actual usage from the CPU subsystem.

The governor is written to be agnostic of the actual counters used to
determine the load. On Krait based CPUs, the governor uses the Krait L2
PM counters which can conflict with certain profiling tools.

The Krait L2 performance monitor counters have the capability to count
different types of requests going to the L2 cache. They also have the
capability to raise interrupts when they overflow. This driver uses
those counters to determine the true usage of L2 from the Krait
processor subsystem and then recommends L2 frequency based on the
measured values and the following tunable parameters.

The driver provides various tunables that allow it to be tuned more in
favor of power or performance:

- cycles_per_high_req: The no. of cache clock cycles that are necessary
  to efficiently process a high-work request to the cache. A higher
  value means higher power and potentially higher performance. A lower
  value means lower power and potentially lower performance.

- cycles_per_med_req: The no. of cache clock cycles that are necessary
  to efficiently process a medium-work request to the cache. A higher
  value means higher power and potentially higher performance. A lower
  value means lower power and potentially lower performance.

- polling_ms: The sampling period in milliseconds. This only affects the
  sampling period when cache use is ramping down or is increasing very
  slowly (See tolerance_mrps).

- min_busy: The minimum percentage of time the cache should be busy.
  This is also applied as a lower bound to the measured busy percentage
  before it's used in calculations. This has to be lower than or equal
  to max_busy. Lower values will make the scaling more aggressive.

- max_busy: The maximum percentage of time the cache should be busy.
  This is also applied as an upper bound to the measured busy percentage
  before it's used in calculations. This has to be greater than or equal
  to min_busy. Lower values will make the scaling more aggressive.

- tolerance_mrps: The minimum increase (in millions of requests per
  second) in cache requests, compared to previous sample, that will
  trigger an IRQ to immediately re-evaluate the cache frequency.

- decay_rate: The parameter controls the rate at which the history is
  forgotten when ramping down. This is expressed as a percentage of
  history to be forgotten. So 100% means ignore history, 0% means never
  forget the historical max. The default 90% means forget 90% of history
  each time.

- guard_band_mhz: This is a margin that's added to the computed cache
  frequency to account for the time it takes between the load increasing
  and the governor/device finishes ramping up the cache frequency.

Change-Id: I918ae178cd3c9d14cb7714d7eb312cbbbb0d977b
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>

Conflicts:
	drivers/devfreq/Kconfig
	drivers/devfreq/Makefile