drivers: mmc: sysfs interface to enable / disable CRC
Enabling software CRCs on the data blocks can be a significant (30%) performance cost, and for other reasons may not always be desired. CRC is a mechanism aiming to prevent data corruption when enabled (reduce the performance around 30%). So if you disable it (improve the performance) but your data can be corrupted. Use it at your risk.

 * ***** SysFs interface :
 *
 * /sys/module/mmc_core/parameters/crc
 *
 * Enable / Disable CRC
 *
 * echo N > /sys/module/mmc_core/parameters/crc (Disabled) or
 * echo 0 > /sys/module/mmc_core/parameters/crc (Disabled)
 *
 * echo Y > /sys/module/mmc_core/parameters/crc (Enlabled) or
 * echo 1 > /sys/module/mmc_core/parameters/crc (Enabled)
 *
 *
 * ***** (default = Enabled)

Change-Id: Ib9c5cb040ab4f40449a430a1b151c6ab2820cc7f
Signed-off-by: Pafcholini <pafcholini@gmail.com>
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>