block: bfq: default to IOPS mode on SSDs
Based on the similar commit for CFQ.

Disabling slice_idle improves throughput on SSDs.

Previously, we disabled slice_idle based on a prop set on boot. However,
if/when the user switches schedulers, slice_idle gets set to 1 again.
Disabling slice_idle in the kernel ensures that slice_idle is always
disabled.

Change-Id: I84212289240b45967bc0c860fb22f586c1c6f5d5
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>